Import into git
This project is imported into git.
It is the command-line interface to execute the project.

Currently, it is only possible to use a configuration from `definiti.conf`
and to check other projects to know how to use it.
A documentation and a getting started will be created when other projects are stable.