Use Docker and nut to encapsulate the project
Because we do not want users to install another tool,
we provide a Docker container and a nut file to encapsulate the tool.

Also, for performances issues, we integrate all dependencies into the docker image
to avoid users to download them afterward and waiting some times.

This commit does the following:

* Create the `Dockerfile` to generate the Docker image
* Add `nut.yml` as an example of nut file
* Create `dependencies/build.sbt` to include definiti dependencies into the image