Remove the need of nut by creating a simple shell script
Asking developers to use nut can block adoption of Definiti.
We now use a shell script to install necessary elements,
so we can continue with a shell script to use definiti.

The drawback is that windows is not compatible (or usable with bash for windows).

This commit does the following:

* Write the script file to execute definiti
* Remove the nut file
* Rename `writeDefinitiFile` to `writeConfFile` for explicitness